pyproject.toml
[project]
name = "energy-trading-system-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "huggingface-hub>=1.9.0",
    "langchain>=1.2.15",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.6.0",
    "prophet>=1.3.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.56.0",
    "transformers>=5.5.0",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
